Move some types to a common area
