+++ src/types/index.ts
@@ -4,3 +4,22 @@ export interface MediaFile {
   path: string
   type: string
 }
+
+export interface Displayer {
+  video: {
+    play: () => Promise<void>
+    pause: () => void
+    setElapsedRatio: (ratio: number) => void
+  }
+}
+
+export interface VideoState {
+  elapsedTime: number
+  elapsedRatio: number
+  isPaused: boolean
+}
+
+export interface PortalState {
+  resource?: { type: string }
+  video?: VideoState
+}
